# packages/backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API E-billets JO 2024",
    "version": "1.0.0",
    "description": "Documentation interactive de l'API e-billets JO 2024 (auth, users, offers)"
  },
  "servers": [{ "url": "http://localhost:4000/api", "description": "Local dev" }],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token JWT retourné",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Inscription utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["firstName", "lastName", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Utilisateur créé" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}